requirements/crewai.txt
numpy==1.26.4
torch==2.3.1
transformers==4.41.2
requests==2.32.3
crewai==0.36.1
crewai-tools==0.4.8
chromadb==0.4.24
spider_env==0.0.1
pydantic-settings==2.3.4